Refactor, regrouping code to packages, separating what is reusable
